BUGFIX: Detect Content-Type in GcsStorage
When uploading a resource to GCS storage the Content-Type is explicitly
defined in metadata options.

```
'metadata' => [
    'contentType' => $resource->getMediaType()
]
```

However `getMediaType()` is always returning the fallback value
`application/octet-stream` as it's logic is based on using the filename
to determine the Content-Type. At this point the filename consists only
of a sha1 or a random value.

PHP's fileinfo extension has functions to detect a Content-Type based on
file content.

This change uses `mime_content_type()` and `finfo_*()` functions (if these
are installed) to detect the Content-Type.